# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  admin.CreateVideoRequest:
    properties:
      actors:
        items:
          $ref: '#/definitions/admin.VideoActorInput'
        type: array
      category_id:
        minimum: 1
        type: integer
      cover_image:
        maxLength: 500
        type: string
      description:
        maxLength: 10000
        type: string
      director_ids:
        items:
          type: integer
        type: array
      duration:
        minimum: 0
        type: integer
      language:
        maxLength: 50
        type: string
      poster_image:
        maxLength: 500
        type: string
      publish_status:
        enum:
        - 0
        - 1
        type: integer
      region:
        maxLength: 50
        type: string
      release_date:
        maxLength: 64
        type: string
      serial_status:
        enum:
        - 1
        - 2
        - 3
        type: integer
      subtitle:
        maxLength: 255
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
      year:
        maximum: 9999
        minimum: 0
        type: integer
    required:
    - category_id
    - title
    type: object
...
  admin.UpdateVideoRequest:
    properties:
      actors:
        items:
          $ref: '#/definitions/admin.VideoActorInput'
        type: array
      category_id:
        minimum: 1
        type: integer
      cover_image:
        maxLength: 500
        type: string
      description:
        maxLength: 10000
        type: string
      director_ids:
        items:
          type: integer
        type: array
      duration:
        minimum: 0
        type: integer
      language:
        maxLength: 50
        type: string
      poster_image:
        maxLength: 500
        type: string
      publish_status:
        enum:
        - 0
        - 1
        type: integer
      region:
        maxLength: 50
        type: string
      release_date:
        maxLength: 64
        type: string
      serial_status:
        enum:
        - 1
        - 2
        - 3
        type: integer
      subtitle:
        maxLength: 255
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
      year:
        maximum: 9999
        minimum: 0
        type: integer
    type: object
...
  admin.VideoActorInput:
    properties:
      actor_id:
        minimum: 1
        type: integer
    required:
    - actor_id
    type: object
  admin.VideoDetailResponse:
    properties:
      actors:
        items:
          $ref: '#/definitions/dto.NamedItem'
        type: array
      category_id:
        type: integer
      cover:
        type: string
      description:
        type: string
      directors:
        items:
          $ref: '#/definitions/dto.NamedItem'
        type: array
      duration:
        type: integer
      id:
        type: integer
      language:
        type: string
      poster:
        type: string
      publish_status:
        type: integer
      rating:
        type: number
      region:
        type: string
      release_date:
        type: string
      serial_status:
        type: integer
      sources:
        items:
          $ref: '#/definitions/admin.VideoSourceGroup'
        type: array
      subtitle:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.NamedItem'
        type: array
      title:
        type: string
      view_count:
        type: integer
      year:
        type: integer
    type: object
  admin.VideoListItem:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      cover:
        type: string
      created_at:
        type: string
      duration:
        type: integer
      id:
        type: integer
      language:
        type: string
      poster:
        type: string
      publish_status:
        type: integer
      rating:
        type: number
      region:
        type: string
      serial_status:
        type: integer
      subtitle:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.NamedItem'
        type: array
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
      year:
        type: integer
    type: object
  admin.VideoSourceEpisode:
    properties:
      episode:
        type: integer
      format:
        type: string
      id:
        type: integer
      quality:
        type: string
      status:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  admin.VideoSourceGroup:
    properties:
      episodes:
        items:
          $ref: '#/definitions/admin.VideoSourceEpisode'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.NamedItem:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  open.VideoDetailItem:
    properties:
      actors:
        items:
          type: string
        type: array
      category_id:
        type: integer
      cover:
        type: string
      created_at:
        type: string
      description:
        type: string
      directors:
        items:
          type: string
        type: array
      id:
        type: integer
      language:
        type: string
      rating:
        type: number
      region:
        type: string
      release_date:
        type: string
      sources:
        items:
          $ref: '#/definitions/open.VideoSource'
        type: array
      subtitle:
        type: string
      title:
        type: string
      year:
        type: integer
    type: object
  open.VideoListItem:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  open.VideoSource:
    properties:
      episodes:
        items:
          $ref: '#/definitions/open.VideoSourceEpisode'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  open.VideoSourceEpisode:
    properties:
      episode:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      cause:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 小橘TV - 影视站系统
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 小橘TV
  version: "1.0"
paths:
...
  /api/admin/v1/videos:
    get:
      consumes:
      - application/json
      description: 分页获取影视列表
      parameters:
      - in: query
        name: actor_id
        type: integer
      - in: query
        name: category_id
        type: integer
      - in: query
        name: director_id
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: language
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - in: query
        name: publish_status
        type: integer
      - in: query
        name: region
        type: string
      - in: query
        name: tag_id
        type: integer
      - in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.VideoListItem'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 影视列表
      tags:
      - 管理端｜视频管理
    post:
      consumes:
      - application/json
      description: 创建一个新的影视
      parameters:
      - description: 影视参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.VideoDetailResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新建影视
      tags:
      - 管理端｜视频管理
  /api/admin/v1/videos/{id}:
    delete:
      description: 删除指定影视
      parameters:
      - description: 影视ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除影视
      tags:
      - 管理端｜视频管理
    get:
      description: 获取指定影视详情
      parameters:
      - description: 影视ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.VideoDetailResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 影视详情
      tags:
      - 管理端｜视频管理
    put:
      consumes:
      - application/json
      description: 更新指定影视信息
      parameters:
      - description: 影视ID
        in: path
        name: id
        required: true
        type: integer
      - description: 影视参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.VideoDetailResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新影视
      tags:
      - 管理端｜视频管理
...
  /api/open/v1/videos:
    get:
      consumes:
      - application/json
      description: 返回启用的影视列表，每项只包含 id、title、category_id、created_at
      parameters:
      - description: DataRange filters by created_at (today/last1d/last3d/last1w/last1m/all).
          Empty = all.
        enum:
        - today
        - last1d
        - last3d
        - last1w
        - last1m
        - all
        in: query
        name: data_range
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Source filters videos by play source name (exact match). Empty
          = no filter.
        in: query
        maxLength: 10
        name: source
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/open.VideoListItem'
                        type: array
                    type: object
              type: object
      summary: 开放影视列表
      tags:
      - 开放端｜开放资源
  /api/open/v1/videos/detail:
    get:
      consumes:
      - application/json
      description: 支持多个视频 id，返回视频详情数组
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: integer
        maxItems: 50
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/open.VideoDetailItem'
                  type: array
              type: object
      summary: 开放影视详情
      tags:
      - 开放端｜开放资源
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"